Compatible with pytorch0.4.0

+++ pretrainedmodels/models/senet.py
@@ -187,9 +187,9 @@ class SEResNeXtBottleneck(Bottleneck):
     expansion = 4
 
     def __init__(self, inplanes, planes, groups, reduction, stride=1,
-                 downsample=None, base_width=4):
+                 downsample=None, base_width=4.):
         super(SEResNeXtBottleneck, self).__init__()
-        width = math.floor(planes * (base_width / 64)) * groups
+        width = int(math.floor(planes * (base_width / 64)) * groups)
         self.conv1 = nn.Conv2d(inplanes, width, kernel_size=1, bias=False,
                                stride=1)
         self.bn1 = nn.BatchNorm2d(width)
